+++ client/src/components/components/Info.jsx
@@ -5,43 +5,16 @@ import {GET_LASTFM_CHARTS} from "../../helpers/queries/lastFm"
 import AlbumInfoPanel from './AlbumInfoPanel';
 import ArtistInfoPanel from './ArtistInfoPanel';
 
-const Info = () => {
+const Info = ({albums, artists}) => {
   const [artistAlbumToggle, setArtistAlbumToggle] = useState(true)
-  let artists = []
-  let albums = []
-  const GetCharts = (tag) => {
-    const { loading, error, data } = useQuery(GET_LASTFM_CHARTS, {
-      variables: { tag },
-        }
-      ) 
-      if (loading) {return 'Loading...';}
-      if (error) {return `Error! ${error.message}`}
-      return data
-    }
 
     const onClickToggle = () => {
       setArtistAlbumToggle(!artistAlbumToggle)
     }
-
-    const randomizer = (data, count, artistOrAlbum) => {
-      const getRandomInt = (max) => {
-        return Math.floor(Math.random() * max);
-      }
-      let i = 0
-      while( i < count){
-        i+=1
-        const selector = getRandomInt(50)
-        // clear tabsArray here
-        artistOrAlbum.push(data[selector])
-      }
-    }
- 
     let tabs;
     let panels;
-    const data = GetCharts("disco")
-      
-    if(artistAlbumToggle === true && data.album){
-      randomizer(data.album, 5, albums)
+
+    if(artistAlbumToggle === true && albums){
       tabs = albums.map((tab)=> {
         return(
             <Tab>
@@ -65,7 +38,7 @@ const Info = () => {
         }
       )
     }
-    else if(!data.album){
+    else if(!albums){
       tabs = (
         <Tab>
           <p>Sharptooth - Mean Brain</p>
@@ -83,7 +56,6 @@ const Info = () => {
       // Sharptooth when data not found (for now) 
     }
     else{
-      randomizer(data.album, 5, albums)
       tabs = artists.map((tab)=> {
         return(
             <Tab>
