+++ client/src/components/pages/Main.jsx
@@ -3,9 +3,13 @@ import React from 'react'
 import Info from "../components/Info"
 import "../../assets/css/style.css"
 import {GET_WEATHER} from "../../helpers/queries/weather"
+import {GET_LASTFM_CHARTS} from "../../helpers/queries/lastFm"
 import Forecast from "../components/Forecast"
 
 const Main = () => {
+    let artists = []
+    let albums = []
+
     const GetWeather = (lat, lon) => {
         const { loading, error, data } = useQuery(GET_WEATHER, {
             variables: { lat, lon },
@@ -20,8 +24,35 @@ const Main = () => {
             />
         )
     }
+    const GetCharts = (tag) => {
+        const { loading, error, data } = useQuery(GET_LASTFM_CHARTS, {
+          variables: { tag },
+            }
+          ) 
+          if (loading) {return 'Loading...';}
+          if (error) {return `Error! ${error.message}`}
+    
+          randomizer(data.album, 5, albums)
+          randomizer(data.artist, 5, artists)
+    
+          return data
+    }
 
+    const randomizer = (data, count, artistOrAlbum) => {
+        const getRandomInt = (max) => {
+          return Math.floor(Math.random() * max);
+        }
+        let i = 0
+        while( i < count){
+          i+=1
+          const selector = getRandomInt(50)
+          // clear tabsArray here
+          artistOrAlbum.push(data[selector])
+        }
+      }
+  
     const weather = GetWeather(1.5, 1.5)
+    const data = GetCharts("disco")
 
     return (
         <div className="wrapper">
@@ -31,7 +62,7 @@ const Main = () => {
             </div>
             <div id="wizard" role="application" className="wizard clearfix">
                 <div className="content clearfix">
-                    <Info />
+                    <Info albums={albums} artists={artists}/>
                 </div>
             </div>
 		</div>
